[CALCITE-3956] Add an abstract class
